wedge100: Move kernel 6.5 patches to 6.6
Summary:
This diff is a no-op since wedge100 uses kernel 6.0. Still,
kernel 6.5 is likely removed from yocto and so bitbake can't find the
`.bb` file to which the `.bbappend` applies and fails to even start the
build.
Since we seem to standardize on kernel 6.6 on other platforms, move the
patches to 6.6 in preparation to upgrading wedge100 kernel to it (if we
ever do it).

Test Plan: CI

Reviewed By: amithash

Differential Revision:
D68349675

Privacy Context Container: L1188897

fbshipit-source-id: 709512c8c3714249517569f17814561664813e50